add all validation and folders, route and components name changed user to employee
